\name{cmdLaunch}
\alias{cmdLaunch}
\title{sleigh Auxiliary Function}
\description{
  This assumes the sleigh object on the master uses
  a worker script to launch remote workers. This function
  is intended to be called within the worker script, which automates
  the process of launching remote workers.
  See RNWSSleighWorker.sh and RNWSleighWorker.py for example.
}


\keyword{methods}
\keyword{internal}
