\name{defaultManagedServerInfoOptions}
\alias{defaultManagedServerInfoOptions}
\title{Default ManagedServerInfo Options Environment}
\description{
This environment specifies the default options that are used
when creating a new managedServerInfo object.
}
\seealso{\code{\link{managedServerInfo}}}

\details{
The default options can be modified by assigning new values
to this environment.  Those new values will be used when
constructing all new managedServerInfo objects.
}

\examples{
defaultManagedServerInfoOptions$port
}

\keyword{environment}
