\name{defaultNetWorkSpaceOptions}
\alias{defaultNetWorkSpaceOptions}
\title{Default NetWorkSpace Options Environment}
\description{
This environment specifies the default options that are used
when creating a new netWorkSpace object.
}
\seealso{\code{\link{netWorkSpace}}}

\details{
The default options can be modified by assigning new values
to this environment.  Those new values will be used when
constructing all new netWorkSpace objects.
}

\examples{
defaultNetWorkSpaceOptions$serverInfo
}

\keyword{environment}
