\name{defaultNwsServerOptions}
\alias{defaultNwsServerOptions}
\title{Default NwsServer Options Environment}
\description{
This environment specifies the default options that are used
when creating a new nwsServer object.
}
\seealso{\code{\link{nwsServer}}}

\details{
The default options can be modified by assigning new values
to this environment.  Those new values will be used when
constructing all new nwsServer objects.
}

\examples{
defaultNwsServerOptions$serverInfo
}

\keyword{environment}
