\name{defaultServerInfoOptions}
\alias{defaultServerInfoOptions}
\title{Default serverInfo Options Environment}
\description{
This environment specifies the default options that are used
when creating a new serverInfo object.
}
\seealso{\code{\link{serverInfo}}}

\details{
The default options can be modified by assigning new values
to this environment.  Those new values will be used when
constructing all new serverInfo objects.
}

\examples{
defaultServerInfoOptions$port
}

\keyword{environment}
