\name{fixedargHolder}
\title{Class "fixedargHolder"}
\alias{fixedargHolder}
\alias{fixedargHolder-class}
\alias{initialize,fixedargHolder-method}
\description{
  For internal use only.
}
\section{Methods}{
  \describe{
    \item{initialize}{\code{signature(.Object = "fixedargHolder", name)}:
      fixedargHolder class constructor.}
  }
}
