\name{germandata}
\docType{data}
\alias{germandata}
\title{German Credit data}
\description{
This dataset produced by Strathclyde University contains numerical 
attributes converted from the original dataset provided by Prof. Hofmann. 
}

\usage{germandata}
\format{
a table contains 1000 instances and 24 attributes for each instance.  
}

\source{
http://ftp.ics.uci.edu/pub/machine-learning-databases/statlog/german
}

\keyword{datasets}
