\name{netWorkSpaceObject}
\docType{methods}
\alias{netWorkSpaceObject-methods}
\alias{netWorkSpaceObject,sleigh-method}
\alias{netWorkSpaceObject}
\title{sleigh Class Method}
\description{
Return the netWorkSpace object associated with the sleigh.
}

\usage{
\S4method{netWorkSpaceObject}{sleigh}(.Object)
}

\arguments{
\item{.Object}{a sleigh class object}
}

\examples{
\dontrun{
s = sleigh()
netWorkSpaceObject(s)
}
}

\keyword{methods}
