\name{nwsWsName}
\docType{methods}
\alias{nwsWsName}
\alias{nwsWsName-methods}
\alias{nwsWsName,netWorkSpace-method}
\title{Return Name of a netWorkSpace}
\description{
Return name of a netWorkSpace.
}

\usage{
\S4method{nwsWsName}{netWorkSpace}(.Object)
}

\arguments{
\item{.Object}{a netWorkSpace class object}
}

\examples{
\dontrun{
ws = netWorkSpace('nws example')
nwsWsName(ws)
}
}
\keyword{methods}
